🚚 ✨ (UI): Moved all pages tests into new test dir
